refactor TopArtist to be TopArtists in all files
