--- src/use_case/TopArtist/TopArtistInputBoundary.java
@@ -1,5 +1,0 @@
-package use_case.TopArtist;
-
-public interface TopArtistInputBoundary {
-    void execute(TopArtistInputData topArtistInputData);
-}
