--- src/use_case/TopArtist/TopArtistOuputBoundary.java
@@ -1,7 +1,0 @@
-package use_case.TopArtist;
-
-public interface TopArtistOuputBoundary {
-    void prepareSuccessView(TopArtistOutputData user);
-
-    void prepareFailView(String error);
-}
