+++ src/use_case/top_artists/TopArtistsDataAccessInterface.java
@@ -1,7 +1,7 @@
-package use_case.TopArtist;
+package use_case.top_artists;
 
 import entity.Artist;
 
-public interface TopArtistDataAccessInterface {
+public interface TopArtistsDataAccessInterface {
     Artist[] getTopArtists(String id, String timeframe);
 }
