+++ src/use_case/top_artists/TopArtistsInputBoundary.java
@@ -1,0 +1,5 @@
+package use_case.top_artists;
+
+public interface TopArtistsInputBoundary {
+    void execute(TopArtistsInputData topArtistsInputData);
+}
