+++ src/use_case/top_artists/TopArtistsInputData.java
@@ -1,10 +1,10 @@
-package use_case.TopArtist;
+package use_case.top_artists;
 
-public class TopArtistInputData {
+public class TopArtistsInputData {
     private String timeframe;
     private final String id;
 
-    public TopArtistInputData(String timeframe, String id){
+    public TopArtistsInputData(String timeframe, String id){
         this.timeframe = timeframe;
         this.id = id;
     }
