+++ src/use_case/top_artists/TopArtistsOuputBoundary.java
@@ -1,0 +1,7 @@
+package use_case.top_artists;
+
+public interface TopArtistsOuputBoundary {
+    void prepareSuccessView(TopArtistsOutputData user);
+
+    void prepareFailView(String error);
+}
